add backup file

+++ WelcomeToVietnam/Controllers/AdminManageController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
+using System.IO;
 using System.Linq;
 using System.Net;
 using System.Web;
@@ -51,7 +52,7 @@ namespace WelcomeToVietnam.Controllers
             {
                 Session["UserAdmin"] = ad.UserName.ToString();
                 Session["ID"] = ad.ID.ToString();
-                return RedirectToAction("HomePage","AdminManage");
+                return RedirectToAction("HomePage", "AdminManage");
             }
             else
             {
@@ -60,6 +61,12 @@ namespace WelcomeToVietnam.Controllers
             return View();
         }
 
+        public ActionResult Logout()
+        {
+            Session["UserAdmin"] = null;
+            Session["ID"] = null;
+            return RedirectToAction("Login", "AdminManage");
+        }
         public ActionResult HomePage()
         {
             if (Session["UserAdmin"] != null)
@@ -68,18 +75,18 @@ namespace WelcomeToVietnam.Controllers
                 return RedirectToAction("Login", "AdminManage");
         }
         // GET: AdminManage/Details/5
-       
+
 
         // GET: AdminManage/Create
         public ActionResult CreatePlace()
         {
             if (Session["UserAdmin"] != null)
-            { 
+            {
                 ViewBag.Area = new SelectList(db.Area, "AreaTravelling", "AreaTravelling");
                 return View();
             }
             else
-                return RedirectToAction("Login", "AdminManage"); 
+                return RedirectToAction("Login", "AdminManage");
         }
 
         // POST: AdminManage/Create
@@ -89,28 +96,34 @@ namespace WelcomeToVietnam.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult CreatePlace([Bind(Include = "Name,Location,Area,Photos,CurrentStatus,VisitorsPerYear")] Place place)
         {
-            if (ModelState.IsValid)
+            if (Session["UserAdmin"] != null)
             {
-                place.totalRatings = 0;
-                place.Rating = 0;
-                db.Place.Add(place);
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
+                if (ModelState.IsValid)
+                {
+                    HttpPostedFileBase file = Request.Files["imageData"];
+                    string path = Path.Combine(Server.MapPath("~/Photos"), Path.GetFileName(file.FileName));
+                    file.SaveAs(path);
+                    byte[] imageArray = System.IO.File.ReadAllBytes(path);
+                    place.Photos = imageArray;
+                    place.totalRatings = 0;
+                    place.Rating = 0;
+                    db.Place.Add(place);
+                    db.SaveChanges();
+                    return RedirectToAction("IndexPlace");
+                }
 
-            ViewBag.Area = new SelectList(db.Area, "AreaTravelling", "AreaTravelling", place.Area);
-            return View(place);
+                ViewBag.Area = new SelectList(db.Area, "AreaTravelling", "AreaTravelling", place.Area);
+                return View(place);
+            }
+            else
+                return RedirectToAction("Login", "AdminManage");
         }
 
         // GET: AdminManage/Edit/5
-        public ActionResult EditPlace(string id)
+        public ActionResult EditPlace(int id)
         {
 
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Place place = db.Place.Find(id);
+            Place place = db.Place.Where(x => x.ID == id).FirstOrDefault();
             if (place == null)
             {
                 return HttpNotFound();
@@ -132,24 +145,26 @@ namespace WelcomeToVietnam.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult EditPlace([Bind(Include = "Name,Location,Area,Photos,CurrentStatus,VisitorsPerYear")] Place place)
         {
-            if (ModelState.IsValid)
+            if (Session["UserAdmin"] != null)
             {
-                db.Entry(place).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                if (ModelState.IsValid)
+                {
+                    db.Entry(place).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("IndexPlace");
+                }
+                ViewBag.Area = new SelectList(db.Area, "AreaTravelling", "AreaTravelling", place.Area);
+                return View(place);
             }
-            ViewBag.Area = new SelectList(db.Area, "AreaTravelling", "AreaTravelling", place.Area);
-            return View(place);
+            else
+                return RedirectToAction("Login", "AdminManage");
         }
 
         // GET: AdminManage/Delete/5
-        public ActionResult DeletePlace(string id)
+        public ActionResult DeletePlace(int id)
         {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            Place place = db.Place.Find(id);
+
+            Place place = db.Place.Where(x => x.ID == id).FirstOrDefault();
             if (place == null)
             {
                 return HttpNotFound();
@@ -162,14 +177,24 @@ namespace WelcomeToVietnam.Controllers
                 return RedirectToAction("Login", "AdminManage");
         }
 
-        [HttpPost, ActionName("Delete")]
+        [HttpPost, ActionName("DeletePlace")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(string id)
+        public ActionResult DeletePlaceConfirmed(int id)
         {
-            Place place = db.Place.Find(id);
-            db.Place.Remove(place);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            if (Session["UserAdmin"] != null)
+            {
+                Place place = db.Place.Where(x => x.ID == id).FirstOrDefault();
+                List<Hotel> hotels = db.Hotel.Where(x => x.Place == place.Name).ToList();
+
+                foreach (Hotel hotel in hotels)
+                    db.Hotel.Remove(hotel);
+                db.Place.Remove(place);
+
+                db.SaveChanges();
+                return RedirectToAction("IndexPlace");
+            }
+            else
+                return RedirectToAction("Login", "AdminManage");
         }
 
         protected override void Dispose(bool disposing)
@@ -180,5 +205,113 @@ namespace WelcomeToVietnam.Controllers
             }
             base.Dispose(disposing);
         }
+        public ActionResult CreateHotel()
+        {
+            if (Session["UserAdmin"] != null)
+            {
+                ViewBag.Place = new SelectList(db.Place, "Name", "Name");
+                return View();
+            }
+            else
+                return RedirectToAction("Login", "AdminManage");
+        }
+
+        // POST: AdminManage/Create
+        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
+        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult CreateHotel([Bind(Include = "Name,Place,Address,Photos,PhoneNumber,Price")] Hotel hotel)
+        {
+            if (Session["UserAdmin"] != null)
+            {
+                if (ModelState.IsValid)
+                {
+                    HttpPostedFileBase file = Request.Files["imageData"];
+                    string path = Path.Combine(Server.MapPath("~/Photos"), Path.GetFileName(file.FileName));
+                    file.SaveAs(path);
+                    byte[] imageArray = System.IO.File.ReadAllBytes(path);
+                    hotel.Photos = imageArray;
+                    hotel.totalRatings = 0;
+                    hotel.Rating = 0;
+                    db.Hotel.Add(hotel);
+                    db.SaveChanges();
+                    return RedirectToAction("IndexHotel");
+                }
+                return View(hotel);
+            }
+            else
+                return RedirectToAction("Login", "AdminManage");
+        }
+
+        [HttpGet]
+        public ActionResult EditHotel(int id)
+        {
+            Hotel hotel = db.Hotel.Where(x => x.ID == id).FirstOrDefault();
+            if (hotel == null)
+            {
+                return HttpNotFound();
+            }
+            if (Session["UserAdmin"] != null)
+            {
+                return View(hotel);
+            }
+            else
+                return RedirectToAction("Login", "AdminManage");
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult EditHotel([Bind(Include = "Name,Address,PhoneNumber")] Hotel hotel)
+        {
+            if (Session["UserAdmin"] != null)
+            {
+                if (ModelState.IsValid)
+                {
+                    db.Entry(hotel).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("IndexHotel");
+                }
+
+                return View(hotel);
+            }
+            else
+                return RedirectToAction("Login", "AdminManage");
+        }
+
+        public ActionResult DeleteHotel(int id)
+        {
+
+            Hotel hotel = db.Hotel.Where(x => x.ID == id).FirstOrDefault();
+            if (hotel == null)
+            {
+                return HttpNotFound();
+            }
+            if (Session["UserAdmin"] != null)
+            {
+                return View(hotel);
+            }
+            else
+                return RedirectToAction("Login", "AdminManage");
+        }
+
+        [HttpPost, ActionName("DeleteHotel")]
+        [ValidateAntiForgeryToken]
+        public ActionResult DeleteHotelConfirmed(int id)
+        {
+            if (Session["UserAdmin"] != null)
+            {
+                Hotel hotel = db.Hotel.Where(x => x.ID == id).FirstOrDefault();
+
+                db.Hotel.Remove(hotel);
+
+                db.SaveChanges();
+                return RedirectToAction("IndexHotel");
+            }
+            else
+                return RedirectToAction("Login", "AdminManage");
+        }
+
+        
     }
 }
